refactor: share logic between resource and data source
